requirements-dev.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.0
distlib==0.4.0
Django==5.2.8
django-cors-headers==4.9.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
gunicorn==23.0.0
identify==2.6.15
idna==3.11
iniconfig==2.3.0
kombu==5.5.4
librt==0.7.3
mypy==1.19.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
ruff==0.14.9
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
vine==5.1.0
virtualenv==20.35.4
wcwidth==0.2.14
